requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
deep-translator==1.11.4
dj-database-url==3.1.2
Django==4.2.28
django-allauth==65.14.3
django-crispy-forms==2.5
django-jazzmin==3.0.2
django-modeltranslation==0.19.17
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
